Refactor and improve type safety across multiple files
Improved `destroyActor` method in `ActorManager.cpp` for clarity.
Simplified vector magnitude calculation in `Comp_Control.h`.
Ensured correct types with `static_cast<int>` in `Comp_Grid.h` and `Grid.cpp`.
Changed `numParticles` type to `int` in `Comp_Particles.h`.
Removed `srand` call from `Engine.cpp` constructor.
Updated `cellSize` return type in `Grid.h` to `float`.
Refactored `LevelManager` methods for better readability and type safety.
Added new color definition `APP_COLOR_TRANSP` in `Params.h`.
Improved type safety in `StateManager.cpp` and `State_Game` files.
Removed unused methods from `Sys` and `SysManager` files.
Added and initialized random number generation variables in `Sys_BunkerControl` and `Sys_InvaderControl`.
Simplified distance calculation in `Sys_ShockwaveControl.cpp`.

+++ Comp_Grid.h
@@ -18,7 +18,7 @@ public:
 			delete m_grid;
 			m_grid = nullptr;
 		}
-		m_grid = new Grid(m_cellSize, ceil(height / m_cellSize), ceil(width / m_cellSize), position.x, position.y);
+		m_grid = new Grid(m_cellSize, static_cast<int>(ceil(height / m_cellSize)), static_cast<int>(ceil(width / m_cellSize)), position.x, position.y);
 	}
 
 	Grid* getGrid()
